refactor: update device
