+++ src/mutenix/updates/constants.py
@@ -1,0 +1,10 @@
+HEADER_SIZE = 8
+MAX_CHUNK_SIZE = 60 - HEADER_SIZE
+DATA_TRANSFER_SLEEP_TIME = 1
+STATE_CHANGE_SLEEP_TIME = 0.5
+WAIT_FOR_REQUESTS_SLEEP_TIME = STATE_CHANGE_SLEEP_TIME
+HID_REPORT_ID_COMMUNICATION = 1
+HID_REPORT_ID_TRANSFER = 2
+
+HID_COMMAND_PREPARE_UPDATE = 0xE0
+HID_COMMAND_RESET = 0xE1
